Fixes some casting warnings that were treated as errors in eclipse

+++ ros_lfd_lat/src/lat_reproducer.cpp
@@ -97,9 +97,9 @@ int main(int argc, char **argv)
 			std::deque< std::deque< double > > reproducedTrajectory;
 			reproducedTrajectory = lfd.reproduce(objects, trajectoryName, "/home/benny/");
 
-			ROS_INFO("length: %i", reproducedTrajectory.size());
-			for (int i = 0; i < reproducedTrajectory[0].size(); ++i) {
-				for (int j = 0; j < reproducedTrajectory.size(); ++j) {
+			ROS_INFO("length: %i", (int)(reproducedTrajectory.size()));
+			for (unsigned int i = 0; i < reproducedTrajectory[0].size(); ++i) {
+				for (unsigned int j = 0; j < reproducedTrajectory.size(); ++j) {
 					std::cout << reproducedTrajectory.at(j).at(i) << "\t";
 				}
 				std::cout << std::endl;
