+++ ros_lfd_lat/src/leatra/dtw_leatra.cc
@@ -11,7 +11,7 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
  
   // Create a list of trajectories in task space
   std::deque< trajectory_lat > TS( JS.begin(), JS.end());
-  for(int i=0; i < TS.size(); i++){
+  for(unsigned int i=0; i < TS.size(); i++){
     TS[i].joint_to_task_space();
   }
   
@@ -23,7 +23,7 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
   // preparing the sequences, that should be warped (for each trajectory one sequence):
   std::deque< std::deque< double > > xy;
 
-  for(int i = 0; i < TS.size(); i++){
+  for(unsigned int i = 0; i < TS.size(); i++){
     xy.push_back( TS[i].get_euclide() );
   }
   
@@ -35,9 +35,9 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
   //     0-4  1-5
   //     0-5
   //
-  for(int i = 0; i < TS.size() - 1; i++){
+  for(unsigned int i = 0; i < TS.size() - 1; i++){
     std::deque< warp > w_row;
-    for(int j = i + 1; j < TS.size(); j++){
+    for(unsigned int j = i + 1; j < TS.size(); j++){
       warp tmp;
       tmp.set_x(xy[i]);
       tmp.set_y(xy[j]);
@@ -57,8 +57,8 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
   int last_i = 0;  // also looking for the most expensive path
   int last_j = 0;
 
-  for(int i=0; i < W.size(); i++){
-    for(int j=0; j < W[i].size(); j++){
+  for(unsigned int i=0; i < W.size(); i++){
+    for(unsigned int j=0; j < W[i].size(); j++){
       if(W[first_i][first_j].get_path_costs() > W[i][j].get_path_costs()){
         first_i = i;
         first_j = j;
@@ -79,8 +79,8 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
   int second_j = last_j;
 
   // Looking for the second cheapest among the trajectory t1
-  if(W.size() > t1){
-    for(int i=0; i < W[t1].size(); i++){
+  if((int)(W.size()) > t1){
+    for(int i=0; i < (int)(W[t1].size()); i++){
       if(W[t1][i].get_path_costs() <= W[second_i][second_j].get_path_costs()){
         if(! (t1 == first_i && i == first_j)){
           second_i = t1;
@@ -90,7 +90,7 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
     }
   }
 
-  for(int i=0; i < W.size(); i++){
+  for(int i=0; i < (int)(W.size()); i++){
     if(t1 - (i+1) > -1){
       if(W[i][t1 - (i + 1)].get_path_costs() <= W[second_i][second_j].get_path_costs()){
         if(! (i == first_i && (t1-(i + 1)) == first_j)){
@@ -102,8 +102,8 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
   }
 
   // Looking for the second cheapest among the trajectory t2
-  if(W.size() > t2){
-    for(int i=0; i < W[t2].size(); i++){
+  if((int)(W.size()) > t2){
+    for(int i=0; i < (int)(W[t2].size()); i++){
       if(W[t2][i].get_path_costs() <= W[second_i][second_j].get_path_costs()){
         if(! (t2 == first_i && i == first_j)){
           second_i = t2;
@@ -113,7 +113,7 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
     }
   }
 
-  for(int i=0; i < W.size(); i++){
+  for(int i=0; i < (int)(W.size()); i++){
     if(t2 - (i+1) > -1){
       if(W[i][t2 - (i + 1)].get_path_costs() <= W[second_i][second_j].get_path_costs()){
         if(! (i == first_i && (t2-(i + 1)) == first_j)){
@@ -151,13 +151,13 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
  std::cout << "Normalizing for Trajectory " << t << std::endl << std::endl;
  
   // Normalizing all paths on t:
-  if(W.size() > t){
-    for(int i=0; i < W[t].size(); i++){
+  if((int)(W.size()) > t){
+    for(unsigned int i=0; i < W[t].size(); i++){
       W[t][i].normalize_path_for_x();
       std::cout<< "NORMALIZING for x: W[" << t << "][" << i << "]" << std::endl;
     }
   }
-  for(int i=0; i < W.size(); i++){
+  for(int i=0; i < (int)(W.size()); i++){
     if(t - (i+1) > -1){
       W[i][t - (i + 1)].normalize_path_for_y();
       std::cout<< "NORMALIZING for y: W[" << i << "][" << t - (i + 1) << "]" << std::endl;
@@ -178,8 +178,8 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
     master_path.push_back(W[t][0].get_path_x());   
     std::cout<< "path number " << t << " from W[" << t << "][" << 0 << "]" << " size = " << W[t][0].get_path_x().size() << std::endl;
   }
-  if(t < W.size()){
-    for(int i = 0; i < W[t].size(); i++){
+  if(t < (int)(W.size())){
+    for(unsigned int i = 0; i < W[t].size(); i++){
       master_path.push_back(W[t][i].get_path_y());
       std::cout<< "path number " << i+t+1 << " from W[" << t << "][" << i << "]" << " size = " << W[t][i].get_path_y().size() << std::endl;
     }
@@ -187,7 +187,7 @@ std::deque< trajectory_lat > warp_leatra::warp_in_task_space(std::deque< traject
    
   // apply warping path to all trajectories in JS
   
-  for(int i=0; i < JS.size(); i++){
+  for(unsigned int i=0; i < JS.size(); i++){
     JS[i].create_this_sequence( master_path[i] );
   }
   
@@ -222,7 +222,7 @@ bool warp_leatra::warp_trajectories(std::deque< trajectory_lat > *tra){
   std::deque< tuple > path = W.get_path();
 
   std::deque< double > rowA, rowB ;
-  for(int i=0; i < path.size(); i++){
+  for(unsigned int i=0; i < path.size(); i++){
     rowA.push_back(x[path[i].x]);
     rowB.push_back(y[path[i].y]);
   }
@@ -235,7 +235,7 @@ bool warp_leatra::warp_trajectories(std::deque< trajectory_lat > *tra){
   ndmap map;
   map.set_name("D5");
   std::deque< std::deque< double> > D = W.get_D();
-  for(int i = 0; i < path.size(); i++){
+  for(unsigned int i = 0; i < path.size(); i++){
     D[path[i].y][path[i].x] = 0.3;
   /* if((path[i].y + 1  <  D.size()) && (path[i].x + 1  <  D[path[i].x].size()) && (path[i].x - 1 > 0) && (path[i].y - 1 > 0)){
       D[path[i].y - 1][path[i].x - 1] = 0.3;
@@ -255,7 +255,7 @@ bool warp_leatra::warp_trajectories(std::deque< trajectory_lat > *tra){
   map.set_name("D_acc5");
 
   std::deque< std::deque< double> > D_acc = W.get_D_acc();
-  for(int i = 0; i < path.size(); i++){
+  for(unsigned int i = 0; i < path.size(); i++){
     D_acc[path[i].y][path[i].x] = 12.0;
  /*   if((path[i].y + 1  <  D_acc.size()) && (path[i].x + 1  <  D_acc[path[i].x].size()) && (path[i].x - 1 > 0) && (path[i].y - 1 > 0)){
       D_acc[path[i].y - 1][path[i].x - 1] = 12.0;
